se corriege error en guardado de imagen en bd por quinta vez en un dia y se crea respaldo de bd

+++ controladores/barcos.controlador.php
@@ -164,57 +164,52 @@ class ControladorBarcos{
 
 	}
 
-
 	/*=============================================
 	EDITAR BARCO
+
 	=============================================*/
 
 	static public function ctrEditarBarco(){
 
-		if(isset($_POST["editarNombreBarco"])){
+			if(isset($_POST["editarNombreBarco"])){
 
 			if(preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["editarDescripcion"]) &&
 			preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["editarNombreBarco"]) &&
 			preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["editarCompania"])){
-
-		   		/*=============================================
+				/*=============================================
 				VALIDAR IMAGEN
 				=============================================*/
 
-			   $ruta = $_POST["imagenActual"];
+				$ruta = $_POST["imagenActual"];
 
-				if(isset($_FILES["editarImagen"]["tmp_name"]) && !empty($_FILES["editarImagen"]["tmp_name"])){
+				if(isset($_FILES["nuevaImagenBarco"]["tmp_name"]) && !empty($_FILES["nuevaImagenBarco"]["tmp_name"])){
 
-					list($ancho, $alto) = getimagesize($_FILES["editarImagen"]["tmp_name"]);
+					list($ancho, $alto) = getimagesize($_FILES["nuevaImagenBarco"]["tmp_name"]);
 
 					$nuevoAncho = 500;
 					$nuevoAlto = 500;
 
-					/*=============================================
-					CREAMOS EL DIRECTORIO DONDE VAMOS A GUARDAR LA FOTO DEL USUARIO
-					=============================================*/
-
-					$directorio = "vistas/img/productos/".$_POST["editarNombreBarco"];
+					$directorio = "vistas/img/productos/".$_POST["editarCodigoBarco"];
 
 					/*=============================================
 					PRIMERO PREGUNTAMOS SI EXISTE OTRA IMAGEN EN LA BD
 					=============================================*/
 
-					if(!empty($_POST["imagenActual"])){
+					if(!empty($_POST["imagenActual"]) && $_POST["imagenActual"] != "vistas/img/productos/default/anonymous.png"){
 
 						unlink($_POST["imagenActual"]);
 
 					}else{
 
-						mkdir($directorio, 0755);
-
-					}	
+						mkdir($directorio, 0755);	
+					
+					}
 
 					/*=============================================
 					DE ACUERDO AL TIPO DE IMAGEN APLICAMOS LAS FUNCIONES POR DEFECTO DE PHP
 					=============================================*/
 
-					if($_FILES["editarFoto"]["type"] == "image/jpeg"){
+					if($_FILES["nuevaImagenBarco"]["type"] == "image/jpeg"){
 
 						/*=============================================
 						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
@@ -222,9 +217,9 @@ class ControladorBarcos{
 
 						$aleatorio = mt_rand(100,999);
 
-						$ruta = "vistas/img/usuar/".$_POST["editarUsuario"]."/".$aleatorio.".jpg";
+						$ruta = "vistas/img/productos/".$_POST["editarCodigoBarco"]."/".$aleatorio.".jpg";
 
-						$origen = imagecreatefromjpeg($_FILES["editarFoto"]["tmp_name"]);						
+						$origen = imagecreatefromjpeg($_FILES["nuevaImagenBarco"]["tmp_name"]);						
 
 						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
 
@@ -234,7 +229,7 @@ class ControladorBarcos{
 
 					}
 
-					if($_FILES["editarFoto"]["type"] == "image/png"){
+					if($_FILES["nuevaImagenBarco"]["type"] == "image/png"){
 
 						/*=============================================
 						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
@@ -242,9 +237,9 @@ class ControladorBarcos{
 
 						$aleatorio = mt_rand(100,999);
 
-						$ruta = "vistas/img/usuarios/".$_POST["editarUsuario"]."/".$aleatorio.".png";
+						$ruta = "vistas/img/productos/".$_POST["editarCodigoBarco"]."/".$aleatorio.".png";
 
-						$origen = imagecreatefrompng($_FILES["editarFoto"]["tmp_name"]);						
+						$origen = imagecreatefrompng($_FILES["editarImagenBarco"]["tmp_name"]);						
 
 						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
 
@@ -258,8 +253,7 @@ class ControladorBarcos{
 
 				$tabla = "barcos";
 
-				$datos = array("id"=>$_POST["idBarco"],
-								"idCategoria" => $_POST["editarCategoriaBarco"],
+				$datos = array("idCategoria" => $_POST["editarCategoriaBarco"],
 							   "nombre" => $_POST["editarNombreBarco"],
 							   "compania" => $_POST["editarCompania"],
 							   "pasajeros" => $_POST["editarPasajeros"],
@@ -271,28 +265,30 @@ class ControladorBarcos{
 							   "cubiertas" => $_POST["editarCubiertas"],
 							   "largo" => $_POST["editarLargo"],
 							   "ancho" => $_POST["editarAncho"],
-							   "imagen" => $ruta);
+							   "imagen" => $ruta,
+								"codigo" => $_POST["editarCodigoBarco"]);
 
 				$respuesta = ModeloBarcos::mdlEditarBarco($tabla, $datos);
+
 
 				if($respuesta == "ok"){
 
 					echo'<script>
 
-						swal({
-							  type: "success",
-							  title: "El barco ha sido editado correctamente",
-							  showConfirmButton: true,
-							  confirmButtonText: "Cerrar"
-							  }).then(function(result){
-										if (result.value) {
+					swal({
+						  type: "success",
+						  title: "El barco ha sido editado correctamente",
+						  showConfirmButton: true,
+						  confirmButtonText: "Cerrar"
+						  }).then(function(result){
+									if (result.value) {
 
-										window.location = "barcos";
+									window.location = "barcos";
 
-										}
-									})
+									}
+								})
 
-						</script>';
+					</script>';
 
 				}
 
@@ -303,7 +299,7 @@ class ControladorBarcos{
 
 					swal({
 						  type: "error",
-						  title: "¡El barco no puede ir con los campos vacíos o llevar caracteres especiales!",
+						  title: "¡El barco no puede ir vacío o llevar caracteres especiales!",
 						  showConfirmButton: true,
 						  confirmButtonText: "Cerrar"
 						  }).then(function(result){
@@ -315,7 +311,9 @@ class ControladorBarcos{
 						})
 
 			  	</script>';
+
 			}
+
 		}
 
 	}
